post-receive: Set the working directory.
When using git+ssh, PWD is set to $HOME, which causes the hook not
to work when bundle install --deployment has been used to install
the gems. This can be fixed by setting the working directory to the
directory to gems have been installed in, in my case the hooks
directory of the git repository.